refactor(auth): 認証フォームのSCSSを共通化
