+++ src/app/(auth)/authForm.module.scss
@@ -1,6 +1,6 @@
 @use 'styles/variables' as *;
 
-// ログインページのメインコンテナ
+// メインコンテナ
 .container {
   width: 100%;
   max-width: 500px;
@@ -14,7 +14,7 @@
   text-align: center;
 }
 
-// ログインフォーム
+// フォーム
 .form {
   background-color: $color-white;
   padding: 2rem;
