--- src/app/(auth)/signup/page.module.scss
@@ -1,102 +1,0 @@
-@use 'styles/variables' as *;
-
-// 新規登録ページのメインコンテナ
-.container {
-  width: 100%;
-  max-width: 500px;
-  margin: 2rem auto;
-}
-
-// ページタイトル
-.title {
-  font-size: 1.8rem;
-  margin-bottom: 1.5rem;
-  text-align: center;
-}
-
-// 新規登録フォーム
-.form {
-  background-color: $color-white;
-  padding: 2rem;
-  border-radius: 8px;
-  border: 1px solid $color-gray-medium;
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-}
-
-// フィールドセット（デフォルトスタイルをリセット）
-.fieldset {
-  border: none;
-  padding: 0;
-  margin: 0;
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-}
-
-// 各入力項目のコンテナ
-.formGroup {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-// 入力項目のラベル
-.label {
-  font-weight: bold;
-}
-
-// 入力フィールド
-.input {
-  width: 100%;
-  padding: 0.75rem;
-  border: 1px solid $color-gray-medium;
-  border-radius: 4px;
-  font-size: 1rem;
-
-  // フォーカス時のスタイル
-  &:focus {
-    outline: none;
-    border-color: $color-primary;
-    box-shadow: 0 0 0 2px rgba($color-primary, 0.2);
-  }
-}
-
-// 相互リンクのコンテナ
-.linkContainer {
-  margin-top: 1.5rem;
-  text-align: center;
-  font-size: 0.9rem;
-  color: $color-gray-dark;
-}
-
-// リンクテキストのスタイル
-.link {
-  color: $color-primary;
-  font-weight: bold;
-  text-decoration: none;
-
-  &:hover {
-    text-decoration: underline;
-  }
-}
-
-// エラーメッセージのスタイル
-.errorMessage {
-  color: $color-error;
-  margin-top: 0.25rem;
-}
-
-// スクリーンリーダー専用テキスト（視覚的に隠すが読み上げ可能）
-.srOnly {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  padding: 0;
-  margin: -1px;
-  overflow: hidden;
-  clip: rect(0, 0, 0, 0);
-  white-space: nowrap;
-  border-width: 0;
-}
